removing Rofi's useless windowrule
PKGBUILD installs the rofi fork for wayland. Rofi runs on the overlay layer which makes this rule useless.

In addition, since this commit ; https://github.com/hyprwm/Hyprland/commit/ec4bea7901bdb1f36d33354c02e36d7e03b1ac1e 
The version 1 of windowrule has been deleted in favour of the v2 syntax, which means that this rule must be modified or removed because it is simply useless.